Merge all assemblies of task into one
This is done by using ILRepack. This got necessary because in some
cases running the task picked up a wrong version of SharpZipLib
(with problems with restoring paths on Linux) despite the correct
version being next to BuildDependencyTasks.dll. I could reproduce
the problem when MONO_PATH contained a wrong SharpZipLib version.
This change works around the problem.