Implement caching of dependencies (#16)
Downloaded files are cached in a directory below
`$LocalApplicationData/BuildDependency` (which is
`.local/share/BuildDependency` or (if set) `$XDG_DATA_HOME` on Linux
and `C:\Users\%USERNAME%\AppData\Local` on Windows). Before
downloading the file we check the cache and copy the file from there.
Caching can be enabled or disabled by setting a new property
`UseCache` on the Dependencies task.